Update Main to handle bigint input.

+++ src/Main.java
@@ -5,8 +5,8 @@ public class Main {
 	public static void main(String[] args) {
 		//Handle input
 		Scanner scanner = new Scanner(System.in);
-		if(scanner.hasNextInt()){
-			System.out.println(Fib.sequence(scanner.nextInt()));
+		if(scanner.hasNextBigInteger()){
+			System.out.println(Fib.sequence(scanner.nextBigInteger()));
 		}
 		scanner.close();
 	}
